pyproject.toml
[project]
name = "modelrunner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "pptozorux", email = "guilhem.chagper@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
]

[project.scripts]
modelrunner = "modelrunner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
